Added financial year to update_forecast url
