+++ pii-ner-exclude.txt
@@ -2041,3 +2041,6 @@ Project 4
 Project 5
 Staff Welfare
 ner_output_file.txt
+Tests 'Edit Forecast'
+Checks Directorate
+r"\t
